pyproject.toml
[tool.black]
line-length = 100
exclude = '''
/(
  | __pycache__
  | .mypy_cache
  | alembic
  | docs
  | envs
)/
'''


[tool.isort]
profile = "black"


[tool.flake8]
max-line-length = 100
exclude = '''
/(
  | __pycache__
  | .mypy_cache
  | alembic
  | docs
  | envs
)/
'''


[tool.mypy]
warn_redundant_casts = true
no_site_packages = true
no_implicit_optional = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
ignore_missing_imports = true
warn_no_return = true
warn_unreachable = true
strict_equality = true


[tool.pylint.MASTER]
# Use multiple processes to speed up Pylint. Specifying 0 will auto-detect the
# number of processors available to use.
jobs=0
max-line-length=100
load-plugins=[
    "pylint_per_file_ignores",
]


# see: https://github.com/christopherpickering/pylint-per-file-ignores
[tool.pylint-per-file-ignores]
"test_db_manager.py"="missing-function-docstring"
"test_utils.py"="missing-function-docstring"
"/routes/"="missing-function-docstring"
"requests.py"="duplicate-code"
"responses.py"="duplicate-code"
"/routes/responses/__init__.py"="duplicate-code"

[tool.poetry]
name = "Savings Manager"
version = "0.1.0"
description = "Manage your savings on your physical account via the app."
authors = ["'PythBuster' <'pythbuster@gmail.com'>"]
readme = "README.md"
packages = [
    { include = "src" },
]

[tool.poetry.dependencies]
python = "^3.11"
alembic = "^1.13.1"
sqlalchemy = "^2.0.25"
pydantic = "^2.6.1"
dictalchemy3 = "^1.0.0"
pydantic-settings = "^2.1.0"
aiosqlite = "^0.19.0"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.1"
fastapi = "^0.109.2"
pytest-cov = "^4.1.0"
uvicorn = "^0.27.1"


[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"
black = {extras = ["jupyter"], version = "^24.1.1"}
isort = "^5.13.2"
flake8-pyproject = "^1.2.3"
pylint = "^3.0.3"
pytest = "^8.0.0"
pytest-dependency = "^0.6.0"
pytest-ordering = "^0.6"
sphinx = "^7.2.6"
pytest-asyncio = "^0.23.5"
pylint-per-file-ignores = "^1.3.2"
vulture = "^2.11"
jupyter = "^1.0.0"
sphinx-rtd-theme = "^2.0.0"


[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--maxfail=1 --cov=src --cov-report html"
testpaths = [
    "tests",
]

[tool.coverage.run]
omit = [
    # omit this single files
    "src/main.py",
    ]

[tool.vulture]
ignore_decorators = ["@model_validator", "@pytest.fixture"]
paths = ["tests", "src", "docs/vulture_whitelist.def"]
sort_by_size = true
make_whitelist = false


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
